Moved KPIs from normal benchmarks
